requirements.txt
# Core
pydantic>=2.0,<3.0
pyyaml>=6.0
requests>=2.31
numpy>=1.26

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# API
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# ML/NLP
sentence-transformers>=2.3.1
transformers>=4.37.0
torch>=2.1.0
tokenizers>=0.15.0

# Dashboard
streamlit==1.30.0
plotly==5.18.0
pandas>=2.1.4

# Utils
python-dotenv>=1.0.0